+++ DLVN-SDK/src/main/java/com/edoctor/dlvn_sdk/store/AppStore.kt
@@ -10,4 +10,5 @@ object AppStore {
     @SuppressLint("StaticFieldLeak")
     var sdkInstance: EdoctorDlvnSdk? = null
     var activeChannelUrl: String? = null
+    var isAppInForeground: Boolean = false
 }